requirements.txt
django-modeltranslation==0.19.15
